requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-gis==1.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
drf-spectacular==0.27.2